fix export dialog
